requirements.txt
Flask==2.0.1
flask-mysqldb==0.2.0
Werkzeug==2.0.1